move nimbledroid files to one place
