+++ src/nimbledroid/NimbledroidSiteDrilldown.jsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { withErrorBoundary } from '../../vendor/errors';
-import NimbledroidGraph from '../../components/NimbledroidGraph';
-import StatusWidget from '../../components/StatusWidget';
-import { siteMetrics } from '../../utils/nimbledroid';
-import fetchNimbledroidData from '../../utils/nimbledroid/fetchNimbledroidData';
+import { withErrorBoundary } from '../vendor/errors';
+import NimbledroidGraph from './NimbledroidGraph';
+import StatusWidget from '../components/StatusWidget';
+import { siteMetrics } from './config';
+import fetchNimbledroidData from './fetchNimbledroidData';
 
 class NimbledroidSiteDrilldown extends Component {
   state = {
