+++ src/nimbledroid/NimbledroidSummaryTable.jsx
@@ -2,11 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-import { generateSitesTableContent } from '../../utils/nimbledroid';
-import fetchNimbledroidData from '../../utils/nimbledroid/fetchNimbledroidData';
-import { withErrorBoundary } from '../../vendor/errors';
-import SummaryTable from '../../components/SummaryTable';
-import StatusWidget from '../../components/StatusWidget';
+import { generateSitesTableContent } from './config';
+import fetchNimbledroidData from './fetchNimbledroidData';
+import { withErrorBoundary } from '../vendor/errors';
+import SummaryTable from './SummaryTable';
+import StatusWidget from '../components/StatusWidget';
 
 const styles = {
   root: {
