+++ test/containers/NimbledroidProductVersions.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import NimbledroidProductVersions from '../../src/containers/NimbledroidProductVersions';
+import NimbledroidProductVersions from '../../src/nimbledroid/NimbledroidProductVersions';
 
 const nimbledroidData = require('../mocks/nimbledroidData.json');
